requirements.txt
streamlit
groq
python-decouple